third_party/blink/web_tests/fast/table/table-different-overflow-values-expected.txt

This test checks for overflow equality. If overflow-x is visible then overflow-y should be visible as well.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS table1: overflow-x is either not visible or both overflow values are visible.
PASS table2: overflow-x is either not visible or both overflow values are visible.
PASS table3: overflow-x is either not visible or both overflow values are visible.
PASS inline_table1: overflow-x is either not visible or both overflow values are visible.
PASS inline_table2: overflow-x is either not visible or both overflow values are visible.
PASS inline_table3: overflow-x is either not visible or both overflow values are visible.
PASS tbody1: overflow-x is either not visible or both overflow values are visible.
PASS tbody2: overflow-x is either not visible or both overflow values are visible.
PASS tbody3: overflow-x is either not visible or both overflow values are visible.
PASS tr1: overflow-x is either not visible or both overflow values are visible.
PASS tr2: overflow-x is either not visible or both overflow values are visible.
PASS tr3: overflow-x is either not visible or both overflow values are visible.
PASS successfullyParsed is true

TEST COMPLETE
